#39 - refactor: 콘솔로그 삭제 및 타입 분리
