+++ src/types/message/roomType.ts
@@ -1,0 +1,14 @@
+export interface GetRooms {
+  status:  number;
+  message: string;
+  data:    Room[];
+}
+
+export interface Room {
+  roomId:      number;
+  image:       string;
+  nation:      string;
+  nickname:    string;
+  recentChat:  string;
+  createdTime: Date;
+}
